Fix whitelist-without-blacklist behavior
The README's description of how to configure an exclusive whitelist did
not reflect reality: if you *only* configured a whitelist, then anything
not in the whitelist would fall through to the typical blacklist logic,
which by default lets everything through if you didn't actually
configure a blacklist.

The change here fixes the logic so that configuring a whitelist but not
a blacklist implies "blacklist everything not explicitly whitelisted".