Fix init with aws passed in
The [StatsD backend interface][1] has another parameter (a logger, which
properly speaking we should be using here instead of console.log). This
was fine before since it was ignored, but when I added the `aws` param
assuming it was "extra" in production, it was actually the logger, which
obviously does not work. Hooray JavaScript.

[1]: https://github.com/etsy/statsd/blob/master/docs/backend_interface.md